okay, it's working. not quite as fast as i'd like

+++ common/include/renderer.h
@@ -41,7 +41,7 @@ typedef struct {
   float sA, cA;  // sin/cos of A
   float sB, cB;  // sin/cos of B
 
-  float angleA, angleB;
+  int angleA, angleB;
 } badge_renderer_t;
 
 // Initialize renderer
